modified fromdate and todate to long values

+++ lcms/lcms-services/src/main/java/org/egov/lcms/models/CaseSearchCriteria.java
@@ -47,9 +47,9 @@ public class CaseSearchCriteria {
 	
 	private String caseCategory;
 	
-	private Integer fromDate;
+	private Long fromDate;
 	
-	private Integer toDate;
+	private Long toDate;
 	
 
 }
